requirements.txt
# Core dependencies
python-dotenv==1.0.0
openai>=1.50.0
streamlit==1.31.0

# NLP libraries
spacy==3.7.4
nltk==3.8.1

# E2B Sandbox
e2b==0.17.1

# Document parsing (PDF and Word)
PyPDF2==3.0.1
python-docx==1.1.0

# Utilities
pandas==2.2.0
numpy==1.26.4
